Visibility Change
Any user that is not in the staffs collection does not have visibility or access on the IT invoices page
