+++ lib/screens/it_invoice_screen.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:graduation_project/components/it_invoice_request_contanier.dart';
 import 'package:graduation_project/components/my_app_bar.dart';
 import 'package:graduation_project/constants.dart';
@@ -17,38 +15,15 @@ enum Status { pending, rejected, done }
 
 class _ItInvoiceScreenState extends State<ItInvoiceScreen> {
   bool? isChecked = false;
-  bool isStaff = false;
+  // Color? statusColor;
+  // String status = "No Status";
+  // Color circleColor = const Color(0XFFE5E5E5);
 
-  @override
-  void initState() {
-    super.initState();
-    checkIfStaff();
-  }
-
-  Future<void> checkIfStaff() async {
-    try {
-      User? user = FirebaseAuth.instance.currentUser;
-      if (user != null) {
-        String email = user.email!;
-        QuerySnapshot querySnapshot = await FirebaseFirestore.instance
-            .collection('staffs')
-            .where('email', isEqualTo: email)
-            .get();
-        if (querySnapshot.docs.isNotEmpty) {
-          setState(() {
-            isStaff = true;
-          });
-        } else {
-          // Redirect to another screen or show a message
-          Navigator.of(context).pushReplacement(
-            MaterialPageRoute(builder: (context) => NotAuthorizedScreen()),
-          );
-        }
-      }
-    } catch (e) {
-      print('Error checking staff status: $e');
-    }
-  }
+  // void updateContainer(Color newColor) {
+  //   setState(() {
+  //     circleColor = newColor;
+  //   });
+  // }
 
   List<RequestContainer> requests = [
     RequestContainer(),
@@ -73,14 +48,6 @@ class _ItInvoiceScreenState extends State<ItInvoiceScreen> {
 
   @override
   Widget build(BuildContext context) {
-    if (!isStaff) {
-      return Scaffold(
-        body: Center(
-          child: CircularProgressIndicator(),
-        ),
-      );
-    }
-
     return Scaffold(
       appBar: PreferredSize(
         preferredSize:
@@ -323,24 +290,6 @@ class _ItInvoiceScreenState extends State<ItInvoiceScreen> {
             )
           ],
         ),
-      ),
-    );
-  }
-}
-
-class NotAuthorizedScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: PreferredSize(
-        preferredSize: const Size.fromHeight(kToolbarHeight),
-        child: MyAppBar(
-          title: 'Not Authorized',
-          onpressed: () => Navigator.pop(context),
-        ),
-      ),
-      body: Center(
-        child: Text('You are not authorized to view this screen.'),
       ),
     );
   }
